# docs/swagger.yaml
basePath: /
host: localhost:3000
info:
  contact: {}
  description: P2P file sharing with content-addressable storage, collection management,
    versioning, merge/fork/pull.
  title: Peerdrive API
  version: "1.0"
paths:
...
  /actions/fork:
    post:
      consumes:
      - application/json
      description: Copy all entries from a source collection into a new collection
        owned by a different user.
      parameters:
      - description: Fork request
        in: body
        name: body
        required: true
        schema:
          properties:
            collection_name:
              type: string
            source_coll_name:
              type: string
            source_username:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: forked, id, username, collection_name, entries_count
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Source collection not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Local collection already exists
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Fork a collection
      tags:
      - actions
...
  /actions/pull:
    post:
      consumes:
      - application/json
      description: Placeholder for syncing upstream changes from a forked source.
        Currently returns a no-op task.
      parameters:
      - description: Pull request
        in: body
        name: body
        required: true
        schema:
          properties:
            collection_name:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Pull upstream updates (placeholder)
      tags:
      - actions
...
  /files/{hash}:
    delete:
      description: Remove file metadata from database and delete the local file if
        provider_type == "local"
      parameters:
      - description: SHA256 hash
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid hash
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a file by hash
      tags:
      - files
  /files/diff:
    post:
      consumes:
      - application/json
      description: Compare version_entries between two collection versions and return
        added/removed/modified lists
      parameters:
      - description: Version IDs to compare
        in: body
        name: body
        required: true
        schema:
          properties:
            version_a:
              type: integer
            version_b:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: added, removed, modified
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Diff entries between two versions
      tags:
      - files
  /files/register_folder:
    post:
      consumes:
      - application/json
      description: Batch-register every file in the given subdirectory of storage/.
        Non-recursive, skips subdirectories.
      parameters:
      - description: Relative folder path under storage/
        in: body
        name: body
        required: true
        schema:
          properties:
            folder_path:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: registered array
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Folder not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register all files in a folder
      tags:
      - files
  /files/register_local:
    post:
      consumes:
      - application/json
      description: Register an already-existing file in the storage directory. Computes
        its SHA256 and inserts into the database without copying.
      parameters:
      - description: Local path and display filename
        in: body
        name: body
        required: true
        schema:
          properties:
            filename:
              type: string
            path:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: hash, filename, note
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request or file not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a local file
      tags:
      - files
  /files/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file, compute its SHA256 hash, store to disk at storage/{first2}/{hash},
        and register in the database. Deduplicates by hash.
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: hash and filename
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Missing file
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: File already exists
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload a file
      tags:
      - files
  /files/verify/{hash}:
    get:
      description: Look up file metadata (filename, provider, path) by its SHA256
        hash
      parameters:
      - description: SHA256 hash
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: hash, filename, provider, path
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid hash
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify a file by hash
      tags:
      - files
...
swagger: "2.0"